+++ mondrian/src/main/java/mondrian/spi/impl/JdbcStatisticsProvider.java
@@ -4,7 +4,7 @@
 * http://www.eclipse.org/legal/epl-v10.html.
 * You must accept the terms of that agreement to use this software.
 *
-* Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
+* Copyright (c) 2002-2016 Pentaho Corporation..  All rights reserved.
 */
 
 package mondrian.spi.impl;
@@ -26,7 +26,7 @@ import javax.sql.DataSource;
 public class JdbcStatisticsProvider implements StatisticsProvider {
     private static final Logger LOG =
         Logger.getLogger(JdbcStatisticsProvider.class);
-    public int getTableCardinality(
+    public long getTableCardinality(
         Dialect dialect,
         DataSource dataSource,
         String catalog,
@@ -75,7 +75,7 @@ public class JdbcStatisticsProvider implements StatisticsProvider {
         }
     }
 
-    public int getQueryCardinality(
+    public long getQueryCardinality(
         Dialect dialect,
         DataSource dataSource,
         String sql,
@@ -85,7 +85,7 @@ public class JdbcStatisticsProvider implements StatisticsProvider {
         return -1;
     }
 
-    public int getColumnCardinality(
+    public long getColumnCardinality(
         Dialect dialect,
         DataSource dataSource,
         String catalog,
